Pretty significant inroads in vir. started to learn project structure using modules etc.
vir is in a state where it can be used for writing and rendering text.
currently working on common functions for moving the line and cursor.
next job is to separate cursor and line movement completely to implement tabs
then adjust the screen to include a minx,y. then improve the renderer.